@@ -2,8 +2,7 @@ import { Injectable, Output, EventEmitter } from '@angular/core';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 
 import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map'; 
-import 'rxjs/add/operator/catch';
+import { map, catchError } from 'rxjs/operators';
 
 import { IUserLogin } from '../../shared/interfaces';
 
@@ -23,22 +22,26 @@ export class AuthService {
 
     login(userLogin: IUserLogin) : Observable<boolean> {
         return this.http.post<boolean>(this.authUrl + '/login', userLogin)
-                   .map(loggedIn => {
-                       this.isAuthenticated = loggedIn;
-                       this.userAuthChanged(loggedIn);
-                       return loggedIn;
-                   })
-                   .catch(this.handleError);
+            .pipe(
+                map(loggedIn => {
+                    this.isAuthenticated = loggedIn;
+                    this.userAuthChanged(loggedIn);
+                    return loggedIn;
+                }),
+                catchError(this.handleError)
+            );
     }
 
     logout() : Observable<boolean> {
         return this.http.post<boolean>(this.authUrl + '/logout', null)
-                   .map(loggedOut => {
-                       this.isAuthenticated = !loggedOut;
-                       this.userAuthChanged(!loggedOut); //Return loggedIn status
-                       return status;
-                   })
-                   .catch(this.handleError); 
+            .pipe(
+                map(loggedOut => {
+                    this.isAuthenticated = !loggedOut;
+                    this.userAuthChanged(!loggedOut); //Return loggedIn status
+                    return status;
+                }),
+                catchError(this.handleError)
+            );
     }
 
     private handleError(error: HttpErrorResponse) {
@@ -49,7 +52,7 @@ export class AuthService {
           // Use the following instead if using lite-server
           //return Observable.throw(err.text() || 'backend server error');
         }
-        return Observable.throw(error || 'Node.js server error');
+        return Observable.throw(error || 'Server error');
     }
 
 }
