@@ -1,98 +1,96 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 
-//Grab everything with import 'rxjs/Rx';
 import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/throw';
-import { Observer } from 'rxjs/Observer';
-import 'rxjs/add/operator/map'; 
-import 'rxjs/add/operator/catch';
+import { map, catchError } from 'rxjs/operators';
 
 import { ICustomer, IOrder, IState, IPagedResults, IApiResponse } from '../../shared/interfaces';
 
 @Injectable()
 export class DataService {
-  
+
     customersBaseUrl: string = '/api/customers';
     ordersBaseUrl: string = '/api/orders';
     orders: IOrder[];
     states: IState[];
 
     constructor(private http: HttpClient) { }
+
+    getCustomersPage(page: number, pageSize: number): Observable<IPagedResults<ICustomer[]>> {
+        return this.http.get<ICustomer[]>(
+            `${this.customersBaseUrl}/page/${page}/${pageSize}`,
+            { observe: 'response' })
+            .pipe(
+                map(res => {
+                    const totalRecords = +res.headers.get('X-InlineCount');
+                    let customers = res.body as ICustomer[];
+                    this.calculateCustomersOrderTotal(customers);
+                    return {
+                        results: customers,
+                        totalRecords: totalRecords
+                    };
+                }),
+                catchError(this.handleError)
+            );
+    }
     
-    getCustomers() : Observable<ICustomer[]> {
+    getCustomers(): Observable<ICustomer[]> {
         return this.http.get<ICustomer[]>(this.customersBaseUrl)
-                    .map(customers => {
-                        this.calculateCustomersOrderTotal(customers);
-                        return customers;
-                    })
-                    .catch(this.handleError);
+            .pipe(
+                map(customers => {
+                    this.calculateCustomersOrderTotal(customers);
+                    return customers;
+                }),
+                catchError(this.handleError)
+            );
     }
 
-    getCustomersPage(page: number, pageSize: number) : Observable<IPagedResults<ICustomer[]>> {
-        return this.http.get<ICustomer[]>(`${this.customersBaseUrl}/page/${page}/${pageSize}`, {observe: 'response'})
-                   .map(res => {
-                       const totalRecords = +res.headers.get('X-InlineCount');
-                       let customers = res.body as ICustomer[];
-                       this.calculateCustomersOrderTotal(customers);
-                       return {
-                           results: customers,
-                           totalRecords: totalRecords
-                       };
-                   })
-                   .catch(this.handleError);
-    }
-    
-    getCustomer(id: number) : Observable<ICustomer> {
+    getCustomer(id: number): Observable<ICustomer> {
         return this.http.get<ICustomer>(this.customersBaseUrl + '/' + id)
-                   .map(customer => {
-                       this.calculateCustomersOrderTotal([customer]);
-                       return customer;
-                   })
-                   .catch(this.handleError);
+            .pipe(
+                map(customer => {
+                    this.calculateCustomersOrderTotal([customer]);
+                    return customer;
+                }),
+                catchError(this.handleError)
+            );
     }
 
-    insertCustomer(customer: ICustomer) : Observable<ICustomer> {
+    insertCustomer(customer: ICustomer): Observable<ICustomer> {
         return this.http.post<ICustomer>(this.customersBaseUrl, customer)
-                   .catch(this.handleError);
-    }
-    
-    updateCustomer(customer: ICustomer) : Observable<boolean> {
-        return this.http.put<IApiResponse>(this.customersBaseUrl + '/' + customer.id, customer)
-                   .map(res => res.status)           
-                   .catch(this.handleError);  
+            .pipe(catchError(this.handleError))
     }
 
-    deleteCustomer(id: number) : Observable<boolean> {
-        return this.http.delete<IApiResponse>(this.customersBaseUrl + '/' + id)
-                   .map(res => res.status)
-                   .catch(this.handleError);
+    updateCustomer(customer: ICustomer): Observable<boolean> {
+        return this.http.put<IApiResponse>(this.customersBaseUrl + '/' + customer.id, customer)
+            .pipe(
+                map(res => res.status),
+                catchError(this.handleError)
+            );
     }
-    
+
+    deleteCustomer(id: number): Observable<boolean> {
+        return this.http.delete<IApiResponse>(this.customersBaseUrl + '/' + id)
+            .pipe(
+                map(res => res.status),
+                catchError(this.handleError)
+            );
+    }
+
     getStates(): Observable<IState[]> {
         return this.http.get<IState[]>('/api/states')
-                   .catch(this.handleError); 
-    }
-    
-    private handleError(error: HttpErrorResponse) {
-        console.error('server error:', error); 
-        if (error.error instanceof Error) {
-          let errMessage = error.error.message;
-          return Observable.throw(errMessage);
-          // Use the following instead if using lite-server
-          //return Observable.throw(err.text() || 'backend server error');
-        }
-        return Observable.throw(error || 'Node.js server error');
+            .pipe(catchError(this.handleError));
     }
 
-    //Not using now but leaving since they show how to create
-    //and work with custom observables
-       
-    createObservable(data: any) : Observable<any> {
-        return Observable.create((observer: Observer<any>) => {
-            observer.next(data);
-            observer.complete();
-        });
+    private handleError(error: HttpErrorResponse) {
+        console.error('server error:', error);
+        if (error.error instanceof Error) {
+            let errMessage = error.error.message;
+            return Observable.throw(errMessage);
+            // Use the following instead if using lite-server
+            //return Observable.throw(err.text() || 'backend server error');
+        }
+        return Observable.throw(error || 'Node.js server error');
     }
 
     calculateCustomersOrderTotal(customers: ICustomer[]) {
@@ -106,5 +104,18 @@ export class DataService {
             }
         }
     }
+
+    //Not using now but leaving since they show how to create
+    //and work with custom observables
+
+    //Would need following import added:
+    //import { Observer } from 'rxjs/Observer';
+    
+    // createObservable(data: any): Observable<any> {
+    //     return Observable.create((observer: Observer<any>) => {
+    //         observer.next(data);
+    //         observer.complete();
+    //     });
+    // }
 
 }
